Separated Mnist_input from input_mnsit.rb and Renamed Mnist_input to MNISTLoader
MNIST読み込みクラスの分離とリネーム
