+++ input_mnist.rb
@@ -1,73 +1,30 @@
-require_relative 'idx_loader'
 require_relative 'fast_neurons'
-require 'zlib'
-
-# Mnistの画像データを入力するクラス
-class Mnist_input
-  # 画像のパスを取得
-  def initialize(image_path, label_path)
-    @image_path = image_path
-    @label_path = label_path
-  end
-
-  # 画像読み込み
-  def load_images
-    @images = load_gzipped_idx_file(@image_path)
-  end
-
-  # ラベル読み込み
-  def load_labels
-    @labels = load_gzipped_idx_file(@label_path)
-  end
-
-  # gzippedfileを読み込み
-  def load_gzipped_idx_file(path)
-    file = File.open(path,"rb")
-    stream = Zlib::GzipReader.new(file)
-    IdxLoader.load(stream)
-  end
-
-  # ピクセルのbyte値をfloatに変換
-  def byte_to_float(inputs)
-    return inputs.map{|pixel|
-      pixel/256.0
-    }
-  end
-
-  # mnistのデータをasciiで出力
-  def ascii_print(inputs)
-    inputs = inputs.map {|pixel| pixel*255}
-    output = inputs.each_slice(28).map do |row|
-      row.map do |darkness|
-        darkness < 64 ?  " " : ( darkness < 128 ? "・" : "X" )
-      end.join
-    end.join("\n")
-    puts output
-  end
-
-end
+require_relative 'mnist_loader'
 
 puts "Loading images"
-mnist = Mnist_input.new("assets/t10k-images-idx3-ubyte.gz", "assets/t10k-labels-idx1-ubyte.gz")
+
+# MNISTの読み込み
+mnist = MNISTLoader.new("assets/t10k-images-idx3-ubyte.gz", "assets/t10k-labels-idx1-ubyte.gz")
 images = mnist.load_images
 
 puts "Initializing network"
-nn = FastNeurons::NN.new([784,15,784]) # ネットワークの作成
-#nn.randomize # ネットワークの初期化
-nn.load_network("network.json") # ネットワークの読み込み
 
+nn = FastNeurons::NN.new([784,15,784]) # ネットワークの作成
+
+#nn.randomize # ネットワークの初期化
+
+nn.load_network("network.json") # 学習したネットワークの読み込み
+
+# pixel値を0 ～ 1の連続値に正規化
 imgs = images.map { |image| mnist.byte_to_float(image).flatten }
 
 puts "Runnning..."
 
 # 学習
-#count = 0
-10.times do
+# サンプルとしてオートエンコーダを構築
+1.times do
   imgs.each.with_index do |inputs,index|
-    # expected_img = mnist.byte_to_float(img).flatten
-    #puts "#{expected_img}"
-    #count += 1
-    #break if count >= 100
+
     nn.input(inputs,inputs) # 入力データと教師データの入力
     nn.run(1) # 実行
 
@@ -75,9 +32,8 @@ puts "Runnning..."
     mnist.ascii_print(nn.outputs) # 学習したデータを出力
   end
 end
-
 puts "Understood!"
-nn.save_network("network.json")
+nn.save_network("network.json") # 学習したネットワークを保存
 gets
 
 # 学習後の確認
